chore: 퍼블릭 API 적용 및 비즈니스 로직 훅 분리
