+++ apps/admin/src/features/welcome-message/index.ts
@@ -1,0 +1,1 @@
+export { useWelcomeMessageForm } from './model/useWelcomeMessageForm';
